Modified code to build as both ANSI and Unicode. Fixed up code that used int for sizes and indexes to use size_t. Replaced c-style casts with C++ casts.

+++ Map.hpp
@@ -58,10 +58,10 @@ public:
 	//
 	// Methods.
 	//
-	int  Count() const;
+	size_t Count() const;
 	void RemoveAll();
 
-	void Reserve(int nItems);
+	void Reserve(size_t nItems);
 
 protected:
 	//
@@ -81,20 +81,20 @@ protected:
 	//
 	// Internal methods.
 	//
-	virtual int Hash(const CMapItem& rItem) const;
+	virtual size_t Hash(const CMapItem& rItem) const;
 
 	//
 	// Members.
 	//
-	int			m_iSize;		// The size of the map.
+	size_t		m_iSize;		// The size of the map.
 	CMapItem**	m_pMap;			// The array of map buckets.
-	int			m_iCount;		// The number of items in the map.
+	size_t		m_iCount;		// The number of items in the map.
 
 	// Map size table size.
 	enum { NUM_MAP_SIZES = 15};
 
 	// Array of map sizes.
-	static int s_aiSizes[NUM_MAP_SIZES];
+	static size_t s_aiSizes[NUM_MAP_SIZES];
 
 	// Max expected chain length.
 	enum { MAX_CHAIN_LEN = 4 };
@@ -115,7 +115,7 @@ private:
 *******************************************************************************
 */
 
-inline int CMap::Count() const
+inline size_t CMap::Count() const
 {
 	return m_iCount;
 }
